+++ messages/en.json
@@ -1,0 +1,6 @@
+{
+  "HomePage": {
+    "title": "Hello world!",
+    "about": "Go to the about page"
+  }
+}
