+++ package.json
@@ -12,6 +12,7 @@
     "eslint": "^8.56.0",
     "eslint-config-next": "^14.1.0",
     "next": "^14.1.0",
+    "next-intl": "^3.26.3",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
