+++ src/app/[locale]/about/page.tsx
@@ -5,17 +5,15 @@ import Breadcrumb from "@/components/Common/Breadcrumb";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "About Page | Free Next.js Template for Startup and SaaS",
-  description: "This is About Page for Startup Nextjs Template",
-  // other metadata
+  title: "About SYNC",
 };
 
 const AboutPage = () => {
   return (
     <>
       <Breadcrumb
-        pageName="About Page"
-        description="Lorem ipsum dolor sit amet, consectetur adipiscing elit. In varius eros eget sapien consectetur ultrices. Ut quis dapibus libero."
+        pageName="About SYNC"
+        description="At SYNC, we are passionate about leveraging technology to solve complex business challenges. With a team of skilled professionals and years of industry experience, we deliver cutting-edge IT solutions tailored to meet the unique needs of our clients."
       />
       <AboutSectionOne />
       <AboutSectionTwo />
