+++ src/app/[locale]/blog-details/page.tsx
