+++ src/app/[locale]/blog/page.tsx
