+++ src/app/[locale]/contact/page.tsx
