+++ src/app/providers.tsx
@@ -4,7 +4,7 @@ import { ThemeProvider } from "next-themes";
 
 export function Providers({ children }: { children: React.ReactNode }) {
   return (
-    <ThemeProvider attribute="class" enableSystem={false} defaultTheme="dark">
+    <ThemeProvider defaultTheme="system" enableSystem>
       {children}
     </ThemeProvider>
   );
