+++ src/components/Blog/SingleBlog.tsx
@@ -1,6 +1,6 @@
+import { Link } from "@/routing";
 import { Blog } from "@/types/blog";
 import Image from "next/image";
-import Link from "next/link";
 
 const SingleBlog = ({ blog }: { blog: Blog }) => {
   const { title, image, paragraph, author, tags, publishDate } = blog;
