+++ src/components/Common/Breadcrumb.tsx
@@ -1,4 +1,4 @@
-import Link from "next/link";
+import { Link } from "@/routing";
 
 const Breadcrumb = ({
   pageName,
@@ -27,7 +27,7 @@ const Breadcrumb = ({
                 <ul className="flex items-center md:justify-end">
                   <li className="flex items-center">
                     <Link
-                      href="/"
+                      href="/home"
                       className="pr-1 text-base font-medium text-body-color hover:text-primary"
                     >
                       Home
