+++ src/components/Header/menuData.tsx
@@ -4,7 +4,7 @@ const menuData: Menu[] = [
   {
     id: 1,
     title: "Home",
-    path: "/",
+    path: "/home",
     newTab: false,
   },
   {
