+++ src/components/Hero/index.tsx
@@ -1,4 +1,4 @@
-import Link from "next/link";
+import { Link } from "@/routing";
 
 const Hero = () => {
   return (
