+++ src/components/ScrollToTop/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useState } from "react";
 
 export default function ScrollToTop() {
