+++ src/routing.ts
@@ -1,0 +1,12 @@
+import { defineRouting } from "next-intl/routing";
+import { createSharedPathnamesNavigation } from "next-intl/navigation";
+
+export const routing = defineRouting({
+  locales: ["en"],
+  defaultLocale: "en",
+});
+
+export const { Link, redirect, usePathname, useRouter } =
+  createSharedPathnamesNavigation(routing);
+
+export type LocaleType = (typeof routing)["locales"][number];
